Fix checksum failure with small isos
When the size of an iso was small enough that the fragment size was
smaller than the buffer size it wouldn't calculate or check the checksum
correctly (it would output 19 instead of 20 checksum fragments).

This fixes that by doing 2 things:
 * read the smaller of:
  - remaining bytes
  - fragment size
  - buffer size
 * When reading the last dangling bit of data add it to the checksum
   but do not write a 21st checksum fragment.

This is backwards compatible with the previous code and now you can
checksum any size iso.