testpyisomd5sum.py: Use a consistent iso size
The test script used to use the directory content for the test.
Depending on the circumstances this could vary in size and fail because
isomd5sum does not work on small isos.

This uses part of the reverted test to make a 500k file/iso to use for
testing. Smaller or larger sizes may be set by passing the size in K on
the cmdline.